requirements.txt
googletrans==4.0.0-rc1
pytest==7.0.0
python-dotenv==0.21.0
python-telegram-bot==20.0
pypinyin==0.53.0
requests==2.28.0